refactor: StudyCircleDetailPageをタブ単位で4コンポーネントに分割
474行の巨大ページを127行に縮小し、各タブを独立コンポーネントに抽出:
- CircleRoadmapTab: ロードマップ・ステップ管理
- CircleCheckinTab: チェックイン入力・履歴表示
- CircleRankingTab: ストリークランキング表示
- CircleSettingsTab: メンバー管理・サークル削除
